+++ Sources/Swift/Tools/Batcher/BatcherScope.swift
@@ -7,7 +7,6 @@ protocol BatcherScope {
     var userObject: User? { get }
     func getContextForKey(_ key: String) -> [String: Any]?
     var attributes: [String: Any] { get }
-    var sendDefaultPii: Bool { get }
 
     func applyToItem<Item: BatcherItem, Config: BatcherConfig<Item>, Metadata: BatcherMetadata>(
         _ item: inout Item,
@@ -73,7 +72,7 @@ extension BatcherScope {
     }
 
     private func addUserAttributes(to attributes: inout [String: SentryAttribute], config: any BatcherConfig) {
-        guard self.sendDefaultPii else {
+        guard config.sendDefaultPii else {
             return
         }
         if let userId = userObject?.userId {