+++ Tests/SentryTests/SentryLogBatcherTests.swift
@@ -344,7 +344,7 @@ final class SentryLogBatcherTests: XCTestCase {
         let attributes = capturedLog.attributes
         
         XCTAssertNil(attributes["sentry.release"])
-        XCTAssertNil(attributes["sentry.trace.parent_span_id"])
+        XCTAssertNil(attributes["span_id"])
         XCTAssertEqual(attributes["sentry.sdk.name"]?.value as? String, SentryMeta.sdkName)
         XCTAssertEqual(attributes["sentry.sdk.version"]?.value as? String, SentryMeta.versionString)
         XCTAssertNotNil(attributes["sentry.environment"])
@@ -370,7 +370,7 @@ final class SentryLogBatcherTests: XCTestCase {
     
     func testAddLog_whenSendDefaultPiiTrue_shouldAddUserAttributes() throws {
         // -- Arrange --
-        scope.sendDefaultPii = true
+        options.sendDefaultPii = true
 
         let user = User()
         user.userId = "123"
@@ -398,7 +398,7 @@ final class SentryLogBatcherTests: XCTestCase {
     func testAddLog_whenSendDefaultPiiFalse_shouldNotAddUserAttributes() throws {
         // -- Arrange --
         let installationId = SentryInstallation.id(withCacheDirectoryPath: options.cacheDirectoryPath)
-        scope.sendDefaultPii = false
+        options.sendDefaultPii = false
 
         let user = User()
         user.userId = "123"
@@ -426,7 +426,7 @@ final class SentryLogBatcherTests: XCTestCase {
 
     func testAddLog_whenSendDefaultPiiTrue_shouldNotAddNilUserAttributes() throws {
         // -- Arrange --
-        scope.sendDefaultPii = true
+        options.sendDefaultPii = true
 
         let user = User()
         user.userId = "123"