+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 
 ### Fixes
 
-- Use `sendDefaultPii` and `span_id` for attributes in `BatcherScope` (#7055)
+- Change default attributes of Logs to only include user attributes when `options.sendDefaultPii = true` (#7055)
+- Rename log attribute `sentry.trace.parent_span_id` to `span_id` (#7055)
 
 ## 9.1.0